chore: move Arguments to args.rs
